Use 64 bits index
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@2734 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/hwacc_vmx.h
@@ -871,7 +871,7 @@ success:
  * @param   idxField        VMCS index
  * @param   u64Val          16, 32 or 64 bits value
  */
-DECLASM(int) VMXWriteVMCS64(uint32_t idxField, uint64_t u64Val);
+DECLASM(int) VMXWriteVMCS64(uint64_t idxField, uint64_t u64Val);
 
 /**
  * Executes VMWRITE
@@ -939,7 +939,7 @@ the_end:
  * @param   idxField        VMCS index
  * @param   pData           Ptr to store VM field value
  */
-DECLASM(int) VMXReadVMCS64(uint32_t idxField, uint64_t *pData);
+DECLASM(int) VMXReadVMCS64(uint64_t idxField, uint64_t *pData);
 
 /**
  * Executes VMREAD
